requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
python-magic==0.4.27
pytz==2024.2
spyne @ git+https://github.com/arskom/spyne.git@f105ec2f41495485fef1211fe73394231b3f76e5
Werkzeug==3.1.3
